Changed default coolboy version to 1 in bat file

+++ make_release.bat
@@ -1,6 +1,8 @@
 make clean
 SET OUTDIR=coolboy-multirom-builder
-del /F /S /Q "%OUTDIR%"
+SET OUTFILE=coolboy-multirom-builder.zip
+IF EXIST "%OUTDIR%" del /F /S /Q "%OUTDIR%"
+IF EXIST "%OUTFILE%" del /F /S /Q "%OUTFILE%"
 mkdir "%OUTDIR%"
 mkdir "%OUTDIR%\tools"
 mkdir "%OUTDIR%\demos"
@@ -16,7 +18,10 @@ copy "!build_rom.bat" "%OUTDIR%"
 copy "tools\CoolboyCombiner.exe" "%OUTDIR%\tools"
 copy "tools\TilesConverter.exe" "%OUTDIR%\tools"
 copy "tools\nesasm.exe" "%OUTDIR%\tools"
+copy "tools\*.dll" "%OUTDIR%\tools"
 copy "spec\*.nes" "%OUTDIR%\spec"
 
 copy "games\*.nes" "%OUTDIR%\games"
 copy "demos\Unchained_Nostalgia.nes" "%OUTDIR%\demos"
+
+7z a %OUTFILE% %OUTDIR%
